feat: create sign up layout, no code added.
