+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.MakeMyDayApp"
         tools:targetApi="31">
         <activity
-            android:name=".SignUpActivity"
+            android:name=".signing.SignUpActivity$SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
