+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,6 @@
     <string name="app_name">MakeMyDayApp</string>
     <string name="sign_up">Sign Up</string>
     <string name="sign_in">Sign In</string>
+    <string name="first_name">First Name</string>
+    <string name="last_name">Last Name</string>
 </resources>
